@@ -29,6 +29,6 @@
     "nodeunit": "^0.9.1",
     "serve-static": "^1.10.0",
     "tsd": "^0.6.0-beta.5",
-    "typescript": "^1.5.3"
+    "typescript": "^1.6.0-beta"
   }
 }
